pyproject.toml
[project]
name = "telomere-airflow-provider"
version = "0.0.1"
description = "Apache Airflow provider for Telomere lifecycle tracking"
authors = [{name = "Module Collective LLC", email = "hello@modulecollective.com"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "apache-airflow>=2.5.0,<3.0.0",
    "requests>=2.28.0",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/modulecollective/telomere-airflow-provider"
Documentation = "https://github.com/modulecollective/telomere-airflow-provider#readme"
Repository = "https://github.com/modulecollective/telomere-airflow-provider"
"Bug Tracker" = "https://github.com/modulecollective/telomere-airflow-provider/issues"

[project.entry-points."apache_airflow_provider"]
provider_info = "telomere_provider:get_provider_info"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
telomere_provider = ["provider.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true